requirements.txt
# Core dependencies
requests==2.32.0
pandas==2.2.0
boto3==1.34.0
python-dotenv==1.0.1

# Data validation
pydantic==2.5.0

# Retry logic
tenacity==8.2.3

# Visualization (optional)
matplotlib==3.8.3
seaborn==0.13.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
requests-mock==1.11.0
moto==4.2.0

# Code quality
flake8==6.1.0
mypy==1.7.0
black==23.12.0

# Logging
python-json-logger==2.0.7
